+++ platform/lint.xml
@@ -4,4 +4,13 @@
     <issue id="Override">
         <ignore path="**/org/conscrypt/OpenSSLX509Certificate.java" />
     </issue>
+
+    <!-- These won't actually be included in the platform build. -->
+    <issue id="Assert">
+        <ignore path="**/org/conscrypt/Platform.java" />
+    </issue>
+    <issue id="NewApi">
+        <ignore path="**/org/conscrypt/Java7PlatformUtil.java" />
+        <ignore path="**/org/conscrypt/Java8PlatformUtil.java" />
+    </issue>
 </lint>
